+++ libs/ajax-action/post-actions/src/ajax-post-action-redirect/ajax-post-action-redirect.service.ts
@@ -1,0 +1,22 @@
+import { Inject, Injectable } from '@angular/core';
+import { InjectionTokenType } from '@spryker/utils';
+
+import { WindowToken } from './tokens';
+import { AjaxPostActionRedirect } from './types';
+
+/**
+ * Changes window location
+ */
+@Injectable({
+  providedIn: 'root',
+})
+export class AjaxPostActionRedirectService {
+  constructor(
+    @Inject(WindowToken)
+    private windowToken: InjectionTokenType<typeof WindowToken>,
+  ) {}
+
+  handleAction(action: AjaxPostActionRedirect): void {
+    this.windowToken.location.href = action.url;
+  }
+}