+++ libs/ajax-action/post-actions/src/ajax-post-action-redirect/index.ts
@@ -1,0 +1,3 @@
+export * from './ajax-post-action-redirect.service';
+export * from './tokens';
+export * from './types';