+++ libs/notification/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-next.8](https://github.com/spryker/ui-components/compare/@spryker/notification@0.1.0-next.7...@spryker/notification@0.1.0-next.8) (2020-06-09)
+
+**Note:** Version bump only for package @spryker/notification
+
+
+
+
+
 # [0.1.0-next.7](https://github.com/spryker/ui-components/compare/@spryker/notification@0.1.0-next.6...@spryker/notification@0.1.0-next.7) (2020-06-04)