+++ libs/styles/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-next.61](https://github.com/spryker/ui-components/compare/@spryker/styles@0.1.0-next.60...@spryker/styles@0.1.0-next.61) (2020-06-09)
+
+**Note:** Version bump only for package @spryker/styles
+
+
+
+
+
 # [0.1.0-next.60](https://github.com/spryker/ui-components/compare/@spryker/styles@0.1.0-next.59...@spryker/styles@0.1.0-next.60) (2020-06-04)